refactor: change package name by kotlin package rule
