+++ src/main/kotlin/carRacing/InputView.kt
@@ -1,4 +1,4 @@
-package car_racing
+package carRacing
 
 object InputView {
