+++ src/test/kotlin/carRacing/GameTest.kt
@@ -1,4 +1,4 @@
-package car_racing
+package carRacing
 
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
